+++ core/src/main/groovy/com/virtusa/gto/nyql/model/QPagedScript.groovy
@@ -1,0 +1,21 @@
+package com.virtusa.gto.nyql.model
+
+import groovy.transform.CompileStatic
+
+/**
+ * @author iweerarathna
+ */
+@CompileStatic
+class QPagedScript extends QScript {
+
+    int pageSize
+
+    QPagedScript(QScript script, int pageSize) {
+        super.id = script.id
+        super.proxy = script.proxy
+        super.qSession = script.qSession
+
+        this.pageSize = pageSize
+    }
+
+}